requirements.txt
Django==1.8.5
MySQL-python==1.2.5
django-guardian==1.3.1
django-notifications-hq==0.8.0
django-userena==1.5.1
pinax-notifications==3.0.1
reportlab==3.2.0
